Renamed modules, fixed annotations() doc
