+++ src/text/mod.rs
@@ -1,6 +1,6 @@
 mod parent_level;
 pub(in crate::text) mod parsers;
-pub mod reader;
+pub mod raw_text_reader;
 mod text_buffer;
 mod text_data_source;
 mod text_value;
